organized files into separate folders
