+++ custom/Helpers/Sardinos.h
@@ -4,18 +4,21 @@
 #include <QGeoCoordinate>
 #include <QList>
 #include <QTime>
+#include <qmath.h>
 #include <algorithm>
 #include <iostream>
 #include <map>
 #include <string>
 #include <unordered_map>
 
-#include <BoundingBox.h>
 #include <LmCdl/VcsiIdentifiedPointOfInterest.h>
 #include <LmCdl/VcsiPointOfInterestProperties.h>
-#include <Waypoint.h>
-#include <WaypointConnector.h>
-#include <qmath.h>
+
+#include "BoundingBox.h"
+#include "Waypoint/Waypoint.h"
+#include "Waypoint/WaypointConnector.h"
+#undef min
+#undef max
 
 class sardinos
 {
@@ -296,6 +299,9 @@ public:
             + atan2(sin(a) * sin(d / R) * cos(lat1),
                     cos(d / R) - sin(lat1) * sin(lat2));
 
-        return QGeoCoordinate(lat2 * (180 / M_PI), lon2 * (180 / M_PI), alt);
+        // Normalize longitude to the range -180 to 180 degrees
+        lon2 = fmod(lon2 + 3 * M_PI, 2 * M_PI) - M_PI;
+
+        return {lat2 * (180 / M_PI), lon2 * (180 / M_PI), alt};
     }
 };
